Make ASHA opt out when last rungs are filled
Why:

When running the trials in the last rungs, it takes a significant amount
of time. During this time, all new trials in lower rungs won't be able
to make up to top rungs, as the final rungs will get completed
meanwhile. This results in significant waste of resources which could be
freed and made available for others.

How:

When penultimate rung is filled, no more trials are promoted nor sampled
at first level. This is because all trials required to select the final
one are already available and no more computations are required beside
completing the current trials in the penultimate rung.