Add arguments low, high, base to Fidelity
Why:

All multi-fidelity algorithms will require these arguments. They do
not have good default values either, so the user is forced to define the
algorithm, whereas passing fidelity(low, high, base) makes the
definition of the algorithm with default values possible
(for the remainder of the arguments, such as `num_brackets` for asha).

How:

Add `low`, `high` and `base` to the fidelity dimension.