Avoid branching in `info`
Why:

When something is different in the environment, the experiment may
branch since we are not using a `View` in `info`. We should rather use a
view but for now they don't have access to instantiated space and
algorithms which are required by `info`.

How:

Use a view and reconfigure it to instantiate the space and algo. This is
not done by default in `View` because they crash if the users scripts in
the experiment are not present on the current FS. At least now this only
happens in `info`, which was already the case anyway since we were not
using a `View`.