Handle race conditions during version increment
Why:

If 2 experiments are created simultaneously and both lead to the same
version increment, one may be created as version=1 and then fail to
resolve the name conflict because version=2 has been registered
meanwhile. This should be resolved by raising an error signaling a race
condition to force the experiment builder to restart the experiment
creation from scratch, as it is done to handle race conditions for
experiment registration during experiment creation.

How:

During conflict detection in experiment.configure, if the
ExperimentNameConflict is not resolved and version number was not
requested (experiment should take last version), then we raise a
RaceCondition error to signal the race condition.