Add functional tests for backward-compatibility
Why:

During release v0.1.6 two bugs have been introduced that made it
impossible to reuse data created with v0.1.5 with the PickledDB
backend. This kind of problems should be automatically detected when
making changes in the develop branch.

How:

For each previous versions (starting with 0.1.3), a db is created using
the old version (0.1.3, 0.1.4, etc) and then all commands are tested in
the current version (test-db, list, status, hunt, etc). The outputs are
not tested thouroughly, this should be covered by the unit-tests. The
tests here only verify that commands run successfully. All tests are run
for each db backend.