Save parser state in DB
Why:

We need the parser state to be able to resume the experiment without the
configuration file. It is also convenient for monitoring scripts where
the configuration file often isn't available locally.
Finally, this is a step in direction of support for python API, as the
prior dict in metadata[prior] will serve as the config used to create
the space instead of metadata[user_args]. This will uniformize how
cmdline and python APIs are handled.

How:

OrionCmdlineParser, CmdlineParser and all Converters now have
`get_state_dict()` and `set_state_dict()` methods so that we can save
the state in metadata and restore the parser without the cmdline and config file.