Refactor experiment builder
Why:

The construction of an experiment is currently very complicated and
coupled with the argument structure coming from the commandline.
By breaking up the parsing of the commandline arguments and the
configuration of the experiment it becomes both easier to build an
experiment and support a new python API in addition to the commandline
API.

How:

Break down the ExperimentBuilder into few public methods, most
importantly `build()`, `build_view()` and `setup_storage()`. In addition
to this, `build_from_args()` and `build_view_from_args()` makes the
bridge between the commandline API and the new experiment builder
functions.

All of `experiment.configure()` is not taking place inside `build()` and
`build_view()`. The mechanisms are currently duplicated to ensure a
smooth migrate, but ultimately the `Experiment` object will solely play
the role of fetching and updating data. Configuration, construction,
branching now all occur in `experiment_builder`.