Add Space definition in DB
Why:

The space definition used to be implicitely stored in
metadata[user_args] when the only API was the commandline one. With the
python API it becomes much more convenient to store the space definition
in the DB explicitely as it is how the user will provide it.

How:

If using the commandline API, parse the user_args, build the space and
save the configuration of the space in DB. If using the python API,
simply save the space in DB directly.