Expose trial.params as a dict
Why:

The list structure is very cumbersome and does not match the interface
provided to users to define space. The params structure as a dict is
more coherent. The ordering of the params is important however for the
hashing of the trial, hence we keep the list as a private attribute and
only expose a dict of {name:value} of the trials, which is more
convenient for the user interface.