Decouple fetch_metadata from command_args process
Why:

We need to fetch the metadata in `build()` in a manner that is agnostic
to the API. The current implementation was expecting the commandline
arguments as inputs.

How:

Function arguments are now user and user_args. If user_args are None, do
nothing, else parse the commandline arguments to fetch script path and
rest of arguments and save in metadata.