Handle broken trials with python API
Why:

If the process exits, the trial status stays to `reserved` and will
eventually be set to `interrupted`.

How:

Register a function to release the remaining trials with pacemakers. If
the process exits normally (no signup signals or sigterm), the function
will be called and all reserved trials will be set to broken.