Move user script inference to OrionCmdlineParser
Why:

The commandline parser is responsible of parsing the commandline (duh)
and thus should also have the responsibility of inferring what the user
script is in the given commandline. The repository check still remains
in resolve_config as this is not specific to commandline parsing. This
enables more complex parsing and reuse of templates to keep original
positions in the commandline (previously it was assumed the script would
be at first position in the commandline).

How:

Infer user script during parsing of commandline and save it as an
attribute of the parser. The `user_args` now keeps the `user_script`
within it, we can't simply pop out the first item anymore. The
`user_script`, if found, is then used outside of the parser to detect
a VCS.