Adapt to new user script inference
Why:

The modification to support more complex inference of user script fixed
a bug that went unnoticed in the functional tests for branching.
The modification of the user script name, which should be detected as a
change in the commandline call, was not detected because it was popped
out of `user_args`. With the modification it was not properly detected,
making plenty of tests fail because they were unfortunately relying on
the bug.

How:

Add --cli-change-type noeffect to let the EVC flow even though there are
changes in the commandline calls (the bug was implicitly doing this).

The test for --non-monitoring-arg was not suppose to lead to any version
change, because the commandline should not be detected as different.
Thus, in this case we remove --cli-change-type noeffect.