Add more helper functions for cmdline client
Some users need a way to signal an interruption from within the script.
While KeyboardInterrupt in the shell where orion's worker is running
would catch the signal properly and interrupt the trial, raising
KeyboardInterrupt from within the script would only lead to a completed
script with error code > 0, hence the trial would be set to broken.

This commit adds a new helper function for this, `interrupt_trial`,
which terminates the process with a special code defined by Oríon, so
that the worker can detect that this is not a broken trial, but
an interrupted one.

Additional helper functions are added to simplify usage of Oríon:
`report_objective` and `report_bad_trials`.