Make global, local and cmdline args coherent
Why:

So far some arguments were supported either in global, local or cmdline,
sometimes in the same format, sometimes in different format. This
was horribly confusing for users.

How:

Force the local config to use the same format as global config by using
the latter as a reference point while parsing local config. The cmdline
arguments are processed and turned into a dictionary following the
structure of the global config.