Add max_trials to algorithm.is_done
Why:

We need to be able to discern when experiment is completed and no more
trials should be executed, and when algorithm is completed and no more
trials should be created. With current experiment.is_done, we sample
more than `max_trials`. Adding `max_trials` to termination criterion of
algo makes it possible for the algo to flag when it should not sample
new points, but still experiment.is_done is False because there is
pending trials.

How:

Add max_trials to algo in `Producer`. This is somewhat hacky but
`max_trials` is tightly integrated to `Experiment` so moving it to
algorithms would be complicated. Maybe something we should consider
doing later on.

In experiment.is_done, it now return True if number of completed trials
is greater than or equal to `max_trials`, or if `algorithm.is_done` and
there is no more pending trials.