Handle timeout in DBs (PickledDB in particular)
Why:

When there are timeouts, the errors are not handled and
python leaves with scary stack-traces in terminal for users.
There is typical issues causing timeouts in PickledDB, the error message
should explain these to the users so they learn what to do to fix it.

Note:

They timeouts never occurred for MongoDB during the stress test and it
is not clear what message should be provided to users, given there is no
obvious root problems as of now. So I leave this as empty reraise, and
do not test...