Fix --debug
Why:

Since the refactoring of experiment builder the debug mode was broken.
This was missed because there was no tests for this in particular.

How:

Add `debug` to global config and move `debug` handling in
`setup_storage` so that whenever `debug` is True, the storage
configuration is overridden with Legacy:EphemeralDB.