Remove useless deprecation warnings
Why:

There was still access to deprecated options in the `core` which were
triggering deprecation warnings event though the user was not using
any of these.

How:

Adjust access to config to use the proper configuration values
and update yaml reading to transfer deprecated settings to proper ones
with warning for users.