Ignore test files for code coverage
Why:

We are not planning to tests the tests, hence it does not make sense to
measure the coverage of the tests on the test code. I would be in favor
of ignoring these files as this often leads to misleading coverage
changes in PRs if they slightly increase code coverage but add
significantly more test code.