Add command orion db rm
Why:

Deleting experiments and trials is a common need expressed by users.

How:

Add a command to delete experiments based on name and version number,
as well as trials based on status.

Methods `delete_experiment` and `delete_trials` are added to the storage
interface and implemented for the Legacy backend. For now the Track
backend does not support it.

To avoid coupling the experiment builder, the EVC and the storagte, the method
`delete_experiment` removes the given experiment only, not the children.
The implementation of the hierarchical deletion is done in the CLI
module (orion.core.cli.db.rm). The same applies for deletion of trials.