Print help for empty cmd calls when invalid
Why:

Some commands such as `orion hunt` cannot be called without any
arguments. The most helpful thing to do in such case is to print the
help message instead of an error.

How:

Mark commands like `hunt` with `help_empty` in the base parser so that it
knows that if no arguments are passed help should be printed.