Add orion db set command
Why:

Users need a way to change values of some fields of the trials, namely
the status.

How:

A command orion db set is added. It supports query for a single
attribute (ex: status=completed) and update for single attributes as
well (ex: status=interrupted). So far only `status` and `id` are
supported for queries and `status` for updates. These are likely to be
sufficient for the majority of use cases.

As for orion db rm, the updates are recursive in children experiments.