Deprecate init_only and add hunt --init-only
Why:

The command init_only was confusing for many users and --init-only as an
argument to `hunt` is more intuitive. Users tended to look for an option
of hunt to have init only instead of looking for another argument.

Note:

The regression tests for backward compatibility relies heavily on
init_only and will need to be adjusted to keep using it for previous
version but use `hunt --init-only` starting with next release. This
modification should be done in release v0.1.10, otherwise tests will
fail in develop until v0.1.9 is released.