Print user friendly err msg if script inexecutable
Why:

A common error is to try running orion on a script that is not
executable. Orion currently prints the whole stack trace to user which
is unclear and somehow scary for users. We should give a clear error
message instead.

How:

Catch PermissionError in Popen and then raise a custom
InexecutableUserScript with a clear error message.
Add InexecutableUserScript to the list of errors that are catched by the
base argparse to print user friendly error messages.