Return onehot interval instead of categorical
Why:

When the categorical dimension is transformed to one-hot, the interval
returned should be in the space of the one-hot otherwise the algorithm
requesting a real space cannot handle the categories.

How:

Add an interval method to OneHot Transformer.
TransformerDimension will look for an interval method, otherwise use the
default interval of original dimension.
Note that CompositeTransformer push the interval() method recursively if
necessary, so that a one-hot encoding that is part of a composite
transformer can still provide its interval().