Adjust OneHotEncoding interval shape
Why:

When the number of categories is only 2, we rely on (0, 1) encoding with
a threshold of 0.5. When number of categories is greater than 2, we rely
on a one-hot encoding. Previous implementation of interval was only
returning the one-hot encoding version, creating a mismatch with shape
when number of categories is 2.