Extend parser support to missing files
Why:

Current implementation could only handle user script that do not exist
in the local file system. The same problem occurs for missing user
script configuration files, but was not handled. This problem occurs
when user wants to load existing experiments on different machines or
locations (if user relative paths) for monitoring.

How:

In cases where config file does not exist, the parser will only fetch
partial configuration search space from cmdargs but not the
user script config file. This should not be problematic when fetching
for monitoring, as the search space is saved in the DB anyway.