Fix flatten for double refs
Why:

If a dictionary contains to references to the same subdict, the deepcopy
used in flatten would copy the subdict but keep the same double refs. In
other words, deep copy does not make 2 different copies for the to same
refs. This is coherent, but modifications within flatten() makes this
problematic as emptying the dict on one ref, affects the other ref (they
are 2 refs to the same objects).

How:

Making local deep copies solves this issue, by creating two copies
instead of 1 copy with 2 refs.