Support load yaml config for dict items
Why:

Some options (like algorithm configuration) have a dict type. The
load_yaml was failing on such configuration because flatten would break
the structure of the configuration, looking at the subkeys of the item's
dictionary as if they were options in the global configuration.

How:

Turn the load_yaml into a recursive function that is called on each
subconfig. This way dictionary items (like algo conf) are not unpacked
and thus processed properly.