Add mock function to force iterative sampling
Why:

We cannot always rely on iterative sampling as same types do not convey
well to it (like large shape, or categorical dims of strings). Must
tests do not need specific sampling, so iterative seeding is fine. For
others, we can mock the sampler to replace the random values by the
seeds, so that we can easily infer what numbers would be returned for a
given seed and test accordingly.