Fix casting of Integer
Why:

There is a mismatch between the inclusive bound of Integer and
the underlying distribution (which is real) where high is in theory exclusive
but in practice inclusive due to limited precision and rounding.
We cannot use numpy.round() to fix this because otherwise the probability
is changed for lowest and highest possible numbers. For instance a
uniform distribution (0, 10) would have a probability of 0.5 * (1/10)
for 0 instead of (1/10).

How:

In the cast method, rescale the point from (low, high) to
(low, high + 1) before computing the numpy.floor().astype(int).
Note that some distributions have infinite intervals, in which case we
do no rescaling.