Handle inf interval in Quantize
Why:

TransformedDimension.interval() will try to apply the transformation on
the (low, high) interval of the original Real dimension. If these are
infinite, the quantization overflows and get stuck with the largest
negative integer.

How:

Detect any inf in interval and replace it with largest possible integer,
preserving the original sign.