Remove methods useless for algos
The ReshapedSpace will not support as much features as the original
space. One reason for this is that the prior strings cannot be
converted according to the transformations. These prior strings are only
useful for the EVC anyway and are of no use for the algorithms. Since
the transformed space are only visible to algorithms, it makes sense to
simplify its interface.