Add auto master-dev sync after release
Why:

After releases on master we need to sync back master to dev.

How:

The github action will automatically make a PR on develop based on
master. For now I believe we should favor a PR instead of direct
push/merge because there may be a need for modifications in the PR
(like citation change to orion version, or version numbers in
documentation).