Support EVC in all algos
Why:

Hyperband and ASHA would fail if they are passed trials that they did
not sample.

How:

Save sampled ids in algos and only observe points that were previously
sampled, otherwise ignore. This way we could still salvages points from
EVC if the current algo sample points that are in EVC (with same
fidelity), otherwise it ignores the points.