Isolate properly DB in branching tests
Why:

The pytest fixtures in branching functional tests was cleaning the DB
but not replacing the DB singleton. Previously, tests running before the
branching ones would use the same DB name so there would be no apparent
issue, but the benchmark tests are using a different DB name leading to
current issue. The fixture save the experiments in `orion` and then the
test runs with a re-initialized `orion_test` db, making the test
failing.

How:

Push back all `create_db_instance` to fixtures and remove useless ones
in the tests.