Add Read/Write/Execution modes for Experiment
Why:

The experiment view is useful to query the database and make plots, but
sometimes we want to add new trials or make modifications to trials
(like status) without executing the experiments. Making such edit should
not require building the experiment with checks for branching, as there
is no intention in executing trials. There should be
read/write/execution rights instead of just providing a view.

Read can:
Make any calls that only require reading the DB. No writes on DB

Read/Write can:
Make any calls that read or write the DB as long as it is not updating
trial results. Also the producer should not be accessible in Read/Write
mode, the algorithm should not be executed in an unchecked environment.

Read/Write/Exec can:
Do everything.

How:

Remove ExperimentView and add modes directly in Experiment. All methods
in Experiment now check it access rights of the experiment allows them
to execute.