Reset global config at end of test
Why:

The test verifies that the global config is properly used for user
script conflict resolution, but fails to reset at the end. I don't know
why it never caused trouble before, maybe the order of tests changed due
to some changes in this PR. Anyhow, this was a bug in the test and it is
now fixed.