Only use localhost as default for mongodb
Why:

The hostname infered with `socket` was causing issues on MongoDB test
not that is was taking place inside MongoDB.__init__. This made me
realize that the inference was never really tested as all unittests were
using a fixture setting the hostname to localhost. I don't know why we
used this inference instead of simply use localhost. If the hostname is
something different, it should still resolve to localhost as well, so I
do not see a need for another default than localhost.