Add generic test-suite for algorithms.
Why:

The use-cases to test are basically the same across all algorithms.
Adding new generic methods or modifying them makes the maintenance of
the tests tedious while they are very consistent across algorithms.
With a test-suite we ensure to cover all basic use-cases for algorithms
and make the maintenance much easier.

How:

Define a base class with all tests and specialization methods that
subclasses can override to customize part of the tests.

Many methods are decorated as _phases_, which makes them parametrizable
with optimization phases. For instance all these tests will run one in
random search phase for TPE and again for the Bayesian Optimization
phases for the TPE. The same applies for Hyperband and its derivatives
to test first phase of random search, different phases of promotions and
phases of bracket repetitions.