Reimplement ASHA on top of Hyperband
Implementing ASHA on top of Hyperband makes it much leaner and easier to
maintain.

Also, the strategy for the allocation of samples to brackets is turned
into a deterministic one. This makes ASHA easier to test and the
behavior is more predictable. Once the first rungs are full according to
the budget then they cannot allow for more trials anymore, but this
new implementation of ASHA supports repetitions of brackets when the
first ones are filled.