Reuse ExperimentClient.workon for CLI workon
Why:

The workon pipeline was duplicated for the cli and python interface. The
python interface should be reused instead under the hood for the CLI
command.

How:

The consumer has been modified to serve only as a callable function to
execute the users script.
- The handling of interruption is moved to ExperimentClient.workon.
- The fetching of results file is moved from storage to consumer since
it is only relevant for the CLI.
- The pacemaker is removed. The heartbeat is only defined and used
within the ExperimentClient.

Experiment.suggest was incoherent when experiment is done. It would
return None, but the context manager could not work in such case. To
handle this, suggest now raises CompletedExperiment instead of returning
None.