Serialize storage singletion when pickling EXP
There are many calls to `get_storage` in the code base that requires the
singletons to be set. When serializing the experiment client to run
multiple workers in parallel the singletons are not accessible in the other
workers. We need to serialize them and reinstate them when deserializing.

This is horribly messy but is required due to over-reliance on
`get_storage()`. This is a bad design and we should get rid of it. See
issue https://github.com/Epistimio/orion/issues/606.