Fix Factory import order by making pack disc lazy
Why:

Trying to import a class that has the factory as a meta class fails if
the Factory is not already imported. This is because during the creation
of the class it will try to do package discovery, thus trying to import
the class that is being created and that triggers the Factory
__init__... To be clear, this happens is the said class is registered in
the python entry points.

How:

The package discovery should be executed lazily when objects are created
with the Factory, not when the class is created. This also makes it
possible to create new Factory classes at run-time.