Add Executor class
Why:

The parallelism backend requires more setup than what
joblib.parallel_backend offers so we need to build our owns.

How:

The backend is fairly simple and only expose `submit` to register tasks
and `wait` to wait for results of all `future`s.

There is also a context manager to switch the executor of the experiment
temporarily. This may be useful to do short distributed optimization and
then switch to local parallelism.