Refactor ExperimentClient to use Executor
Why:

The optimization was tied to joblib. We now move to Executor to allow
using multiple backends directly instead of through joblib.

How:

All joblib.delayed() calls are converted to executor.submit() and
joblib.Parallel()() is converted to executor.wait().

The experiment is assigned a default executor if none are passed.
It is possible to use a temporary executor with `client.tmp_executor` so
that the default one is reassigned at end of context. This will prove
useful when some executor temporary wraps the storage, so that out of
the parallel execution we can be sure to access the unwrapped storage.