Add CLI options for Executor backend
Why:

The new Executor backend was only usable in python API so far. This
commit add arguments `--n-workers` and `--executor` plus support in
global/local configs for `worker/executor_configuration`.

Note:

Setting the signal inside the consumer call was problematic because it
should only be set in the main thread. Because of this the signal is now
set in `cli/hunt.py`.