Add enable option for EVC, with default = false
The EVC is a constant source of confusion for users. It should be
disabled by default with warning messages when different versions of
experiments are used. Users who wants using advanced features of the EVC
would still be able to use it by enabling it.

Making it false by default is a breaking change and may cause issues to
user currently using the EVC. Based on discussion with users there does
not seem to have much usage of the EVC so far so this breaking change
should be relatively harmless. Avoiding further confusion by making it
disabled by default is worth the breaking change.