Allow branching when parent script isn't available
Why:

Suppose we want to branch from a parent running from a different
computer, or for which we lost the execution script. The branching
should not fail because script is missing, we do not need it because we
wont execute trials from the parent anyway.

How:

Use allow_non_existing_files=True when building the cmdline parser to
compare cmdlines of experiments.