Remove space upgrade from DB upgrade
Why:

The space upgrade relies on local files if the experiment's search space
is defined in a configuration file. Parsing these file during the DB
upgrade can break the DB because all experiments may not be executed on
the same file system and thus some configuration files may not be
present. The space should only be upgraded when the user attempts
running an experiment, in which case the configuration file is available.

The space does not need to be upgraded during DB upgrade anyway, because
experiment built is backward compatible with experiments lacking an
explicit space definition in DB (relying on cmdargs and config file to
define space at run-time).

How:

Remove space upgrade from db upgrade script.