Filter out trials when no default value available
Why:

When a dimension is deleted or added, the adaptor should not transform
them with a default value of None if there was no default values. This
would lead to invalid trials if None is an invalid value of this
dimension.

How:

If the default value is the unique NO_DEFAULT_VALUE object, then the
trials should be filtered out.