Compute cardinality for loguniform with precision
Why:

With loguniform the number of possible values is limited if precision is
used. Cardinality computation should account for this otherwise
algorithms may get stuck in suggest(). It happened to a user with a
prior loguniform(1e-4, 1e-2, precision=2). This gives only 181 possible
values.

How:

If real dimension has precision and prior loguniform, then compute
cardinality.

There is a problem with transformed space however. A linearized
dimension for instance would attempt to compute the cardinality with the
linearized bounds. What matters is the smallest cardinality between the
transformed space and the original space. The only case where
cardinality is smaller in transformed space is when real values are
discretized. Therefore, we only compute cardinality of transformed
dimensions if transformation lead to integer, otherwise we use the
cardinality of the original dimension.