Fix lost trials of parents
Why:

When running an experiment, the parents may have lost trials that are
stuck to status reserved. If the user cannot run the parents, it must
then use `orion db set` to fix this manually. This should be done
automatically instead.

How:

Loop over the EVC and call `fix_lost_trials` for each experiments. Note
that this can increase significantly the cost of the command for large
EVC.

A ugly hack is used to allow running `fix_lost_trials` on the parents
that are in read mode. It would be great to find a better work around...