Handle properly all types in config during branch
Why:

During resolution of conflicts, the user's script configuration file is
parsed for any marked resolution. This parsing is not handling properly
values that may not be string and causes Oríon to crash.

How:

The values should only be handled if they are string and contain the
value '~' as this is the only way to mark a resolution. They can be
safely ignored otherwise.